# api-gateway/docs/swagger.yaml
definitions:
...
  dto.Category:
    description: Category details
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.CreateCategoryRequest:
    description: A request structure for creating a cateogry
    properties:
      name:
        example: Category Name
        type: string
    required:
    - name
    type: object
  dto.CreateCategoryResponse:
    description: A success message for category creation
    properties:
      message:
        type: string
    type: object
...
  dto.DeleteCategoryResponse:
    description: A success message for category deletion
    properties:
      message:
        type: string
    type: object
  dto.ErrorResponse:
    description: Error response structure
    properties:
      message:
        type: string
    type: object
...
  dto.GetCategoriesResponse:
    description: A list of categories
    properties:
      categories:
        items:
          $ref: '#/definitions/dto.Category'
        type: array
    type: object
...
  dto.UpdateCategoryRequest:
    description: A request structure for updating a cateogry
    properties:
      name:
        example: Category Name Updated
        type: string
    required:
    - name
    type: object
  dto.UpdateCategoryResponse:
    description: A success message for category update
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of categories from the gRPC service
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetCategoriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a list of categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category with the provided information
      parameters:
      - description: Category information
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category with the provided information
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update a category with the provided information
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category information
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a category
      tags:
      - Categories
...
swagger: "2.0"